backend/api/requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jwt==1.3.1
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
